+++ CHANGELOG.rst
@@ -6,4 +6,4 @@ Unreleased
 ==========
 * fix: Failing ci and test suite
 * feat: Removed Django 2.0 and 2.1 from CI tests
-* feat: Added file list item actions burger menu
+* feat: Customisable folder list admin file action buttons